User integration with AgoraVoting.

+++ mieli/api/organization.py
@@ -1,34 +1,51 @@
 from identity.models import Organization
 from django.contrib.sites.models import Site
+from django.db import transaction
+from mieli import registry
 import random
 import string
 import copy
 
-def setup_filter(args):
-    cargs = copy.copy(args)
-    for key in args.iterkeys():
+def setup_query(kwargs):
+    query = copy.copy(kwargs)
+    for key in kwargs.iterkeys():
         if key == 'domain' or key == 'name':
-            value = cargs.pop(key)
-            cargs['site__%s' % key] = value
-    return cargs
+            value = query.pop(key)
+            query['site__%s' % key] = value
+    return query
 
-def get(**args):
-    query = setup_filter(args)
+def get(**kwargs):
+    query = setup_query(kwargs)
     try:
         organization = Organization.objects.get(**query)
     except Organization.DoesNotExist:
         organization = None
     return organization
 
+def get_by_username(username):
+    if not '@' in username:
+        raise Exception("invalid username '%s', expected username@organization" % username)
+    domain = username.split('@')[1]
+    query = { 'domain': domain }
+    return get(**query)
+
+@transaction.atomic
 def create(domain, name):
+    if get(domain=domain):
+        raise Exception('domain %s is already in use' % domain)
     site = Site(domain=domain, name=name)
     site.full_clean()
     site.save()
     organization = Organization(site=site)
     organization.full_clean()
     organization.save()
+    registry.invoke('organization_create', organization=organization)
+    return organization
 
-def delete(**args):
-    query = setup_filter(args)
-    organization = get(**query)
+@transaction.atomic
+def delete(**kwargs):
+    organization = get(**kwargs)
+    if organization == None:
+        raise Exception('unknown organization')
+    registry.invoke('organization_delete', organization=organization)
     organization.site.delete()
